Add new LossScaleOptimizerV3 subclassing from the new experimental optimizer.
The existing LossScaleOptimizer does not support the new experimental optimizer in the file keras/optimizer_experimental/optimizer.py, so a new LossScaleOptimizerV3 class is created to support it. LossScaleOptimizerV3 subclasses from the experimental optimizer and wraps instances of the experimental optimizer. LossScaleOptimizerV3 is essentially a copied version of the old LossScaleOptimizer class, with some minor changes to support the interface of the new optimizer.

I copied the old LossScaleOptimizer, since I couldn't find a clean way of factoring out the common code. I tried creating a LossScaleOptimizerBase mixin class with the common code that both LossScaleOptimizer and LossScaleOptimizerV3 would subclass. However, this turned out to be very complicated since the two base optimizers (OptimizerV2 and the experimental Optimizer) have different method names and argument names, so I didn't go with this approach. While the code in LossScaleOptimizer is copied to LossScaleOptimizerV3, there are still many helper functions/classes that the two classes use, so most of the loss scaling-related code is not copied.

PiperOrigin-RevId: 409001050